# institute-management-front/src/app/services/models/admin-update-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AdminUpdateRequest {
  age?: string;
  email?: string;
  gender?: string;
  isApproved?: boolean;
  name?: string;
  role?: string;
}