# institute-management-front/src/app/services/models/course-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CourseResponse {
  createdAt?: string;
  endedAt?: string;
  id?: string;
  name?: string;
  startedAt?: string;
}