# institute-management-front/src/app/services/models/user-search.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserSearch {
  age?: string;
  email?: string;
  gender?: string;
  isApproved?: boolean;
  isProfileCompleted?: boolean;
  isVerified?: boolean;
  name?: string;
  role?: string;
}